Models

In this package, the common interface to advance simulations is to apply functions of the form x = f!(x, Δt::Float64; parameters...) to the current simulation state x, and we call the effects of a single model invocation a simulation segment. The state x has no fixed representation and may even change type within f!, so while f! may modify x in place, the canonical result is given by its return value. The time step Δt ≥ 0.0 indicates the units of simulation time that f! is supposed (but not required) to advance. We refer to functions that adhere to this contract as models, and all of those provided in this package are implemented as (immutable) functors of type <: Model.

Models are of two kinds: dynamics models that represent continuous-time behaviour and thus advance the simulation state along the time axis, and instant models that represent instantaneous adjustments and do not actually advance along the time axis. The former includes all gene regulation models, while the latter, defined as <: Instant, includes various forms of instant adjustments that may be used to construct interventions as well as observation schemes. Besides those provided, it is straightforward to define new Models and state representations and wire them up with the rest of the package; the following describes the interface that needs to be implemented to achieve this.

State

Different models may require different representations of the simulation state to operate on. There is no common supertype of these state representations, but all of these State types must at least carry information about the current simulation time, the values of all modeled dynamic variables at that time, and the current state of a random number generator. For the dynamics models implemented in this package, the state object also accumulates the full simulation trajectories of a simulation segment if their output is required.

This information is accessed using the following methods, which must be implemented for all state types:

GeneRegulatorySystems.Models.t — Function

t(x)

Access the current simulation time of state x.

GeneRegulatorySystems.Models.randomness — Function

randomness(x)

Access the random number generator instance of state x.

GeneRegulatorySystems.Models.each_event — Function

each_event(callback::Function, x)

Invoke callback once for each state change event captured in state x.

This will iterate the state in long format, with one event per dimension at the earliest contained timepoint and then, if the trajectory is included in x, every change event ordered by timepoint. The required signature is callback(t::Float64, key::Symbol, value::Int).

(I intend to reimplement this as a Tables.jl rows interface.)

Examples

julia> x = Models.FlatState(counts = Dict(:a => 10));

julia> Models.each_event(println ∘ tuple, x)
(0.0, :a, 10)

julia> Models.each_event(x) do t, key, value
           println((; t, key, value))
       end
(t = 0.0, key = :a, value = 10)

Most Instant models in this package use FlatState as their state representation, which simply contains the current species counts in a Dict{Symbol, Int} (i.e., not nested for subsystems) but does not retain full trajectories. For FlatStates, each_event emits one event for each contained species at the current simulation time.

GeneRegulatorySystems.Models.FlatState — Type

FlatState

Represents species counts at a single time point.

Counts are stored as a flat Dict{Symbol, Int} mapping dimension names to counts. For models that have dimensions associated with subsystems (such as genes in Models.V1), the corresponding dimension names in FlatState are flattened by joining on ".".

For example:

julia> f! = Models.Plumbing.Wait();

julia> g! = Models.Plumbing.adder(Dict(:a => 10));

julia> x = Models.FlatState();

julia> x = f!(x, 100.0);

julia> x = g!(x, 0.0);

julia> Models.t(x)
100.0

julia> Models.each_event(x) do t, key, value
           println((; t, key, value))
       end
(t = 100.0, key = :a, value = 10)

Instantiating models

Although model functors can be created directly like in the preceding example, in practice they will likely be constructed by loading a JSON definition:

GeneRegulatorySystems.Models.parse — Function

parse(
    definition;
    into = "",
    channel = "",
    defaults = load_defaults(),
    others...,
)

Parse a JSON definition to construct a Model.

The definition can be a String or an IO, and keyword arguments will be collected and passed as bindings for the model construction.

Examples

julia> f! = Models.parse("""
           {"{add}": {"a": 10}}
       """)
GeneRegulatorySystems.Models.Plumbing.Adjust(+, Dict(:a => 10))

This is equivalent to the following:

julia> import JSON

julia> source = """
           {"{add}": {"a": 10}}
       """;

julia> definition = JSON.parse(source, dicttype = Dict{Symbol, Any})
Dict{Symbol, Any} with 1 entry:
  Symbol("{add}") => Dict{Symbol, Any}(:a=>10)

julia> f! = Model(definition)
GeneRegulatorySystems.Models.Plumbing.Adjust(+, Dict(:a => 10))

GeneRegulatorySystems.Models.load — Function

load(path::AbstractString; bindings...)

Load a Model from a JSON file.

This is just a convenience wrapper around parse; keyword arguments will be forwarded.

SciML-based jump process models

While the simulation interface in principle allows the definition of arbitrary dynamics, all of the first-iteration gene regulation model templates currently included in this package are ultimately interpreted as sampling trajectories from (discrete-space) pure jump processes. This choice is also reflected in the current implementation of instant adjustment and observation models as well as the command-line wrapper scripts and their shared results format (which records discrete state change events in long format).

Originally built using a custom Gillespie-type integrator, the included regulation models have subsequently been migrated onto the SciML ecosystem, specifically Catalyst.jl and JumpProcesses.jl. This allows using those tools to assemble systems and to sample trajectories with fast off-the-shelf stochastic simulation algorithms such as RSSACR, but also enables further analysis of the intermediate model representations with the various SciML packages.

Common machinery to build and simulate such SciML-based models are defined in the Models.SciML module, which contains JumpModel <: Model{JumpState} and the corresponding JumpState.

GeneRegulatorySystems.Models.SciML.JumpModel — Type

JumpModel <: Model{JumpState}

Represents the stochastic dynamics of applying a JumpProcesses.AbstractAggregatorAlgorithm method to a ModelingToolkit.JumpSystem system with a set of parameters.

Gene regulation models in this package ultimately get compiled to JumpModels.

Specification

In JSON, JumpModels can only be defined indirectly, such as via Models.V1.

Invocation

(f!::JumpModel)(x::JumpState, Δt::Float64; record = false, _...)

Advance the simulation by applying the stochastic dynamics f! to x for Δt time units, realizing a segment of the state trajectory.

x must be compatible with f!, that is, the JumpProcesses.JumpProblem (and corresponding integrator) in x must be compatible with f!.system. This is currently conservatively checked as f! === x.f!. If necessary, users can call adapt!(x, f!) to convert x appropriately.

If record === true, x.integrator will record all jumps, otherwise the trajectory will not be retained (and only the final state will be available). Either way, the recorded trajectory will be initially cleared, so it needs to be extracted before the next invocation of f!.

Unfortunately, JumpProcesses.jl always uses a dense trajectory encoding, so that the recorded trajectory information is highly redundant and needs to be filtered by each_event for output in sparse long format.

GeneRegulatorySystems.Models.SciML.JumpState — Type

JumpState

Contains the prepared JumpProcesses.JumpProblem problem and an associated JumpProcesses.SSAIntegrator integrator to be advanced by a JumpModel.

Since in JumpProcesses.jl the problem and integrator objects are tightly coupled (because they alias various components), JumpState holds both of them. The current time and state as well as the (potentially) recorded trajectory are contained in the integrator, while the random number generator is part of the problem object.

To be compatible with a JumpModel f!, the JumpState's problem (and integrator) must have been constructed from the same f!.system. To check whether this is actually the case, JumpState also contains a reference to the corresponding f!.

For JumpStates, each_event emits events depending on whether the preceding model invocation was instructed to record the trajectory or not. Without recording, each_event behaves similarly to the FlatState case, emitting one event for each contained species at the final simulation time. With recording, each_event emits one event per species at the beginning of the trajectory, and subsequently one event for every recorded change. The species names will be converted to Symbols, replacing potential ₊s (subsystem separators) by .s.

Connecting model invocations

Consecutive invocations of models that operate on different state representations require a state conversion in between. For this purpose, the adapt! function can be used: it transforms a state to the type required for a given Model, and it is predefined for all state-model combinations included in this package.

GeneRegulatorySystems.Models.adapt! — Function

adapt!(x, f!; copy = false)

Convert the simulation state x to a type accepted by the model f!.

If copy is set, x is not modified and the return value independent of it, except for retaining the the same randomness instance. Otherwise, adapt! is allowed to modify or alias parts of x, which should therefore no longer be used.

This design (having a type-unstable state representation) generally assumes that the total number of simulation segments is not too large and that the majority of computation time is spent within the primitive model invocations. A further caveat is that the construction of JumpProcesses.jl JumpProblems and initializing its integrators is quite costly and therefore repeatedly adapt!ing state to JumpModels, for example after executing periodic interventions, can be slow in our current implementation.

Defining a new Model or state type State typically also requires defining new adapt! methods. However, if no specific method is found for a given state-model pair, adapt! defaults to retrying with a FlatState copy of the state. This means that often, defining FlatState(x::State) and adapt!(x::FlatState, f!::Model, _copy) is sufficient (see below).

Model invocations both of instant and dynamics models can be organized in Schedules, which are themselves Models that execute multiple simulation segments when invoked. The scheduling machinery internally uses adapt! to piece the model segments together.

Typically, adapt! may modify or alias its state parameter when preparing the result. But since Schedule also supports trajectory branching, adapt! can be instructed (with its copy parameter) to make sure that the resulting state is independent of the input (besides always aliasing the used random number generator).

The Model interface

GeneRegulatorySystems.Models.Model — Type

Model{State}

Abstract supertype of all models.

Models are functors: subtypes M <: Model{State} must define a call overload method (::M)(x::State, Δt::Float64; arguments...) that advances the simulation. Subtypes indicate which State type they expect to receive, and must accept at least such x.

If State is a newly defined type (that is, specific to M and e.g. not FlatState), that also requires implementing at least the following methods:

each_event(callback::Function, x::State) to extract the state (and, if contained, trajectory) in long format.

t(x::State) to access the current simulation time.

randomness(x::State) to access the contained random number generator.

adapt!(x::FlatState, f!::M, copy::Val) to convert a FlatState to a State and return it. The result must alias x.randomness. If copy is Val(true), the result must otherwise be an independent deep copy of x.

FlatState(x::State) to allow adapt! for subsequent models to fall back on converting to FlatState and retrying if there is no more specific adapt! method defined.

However, implementing adapt! methods for more specific state-model-pairs may allow for more efficient state conversion between model invocations, for example because a copy isn't required or parts of another state type can be reused. adapt! is used by Schedule to convert between the various state representations required by its contained primitive Models, but it may also be called directly.

In addition, a Specifications.constructor(::Val{<:Symbol}) method typically needs to be defined to return a function that instantiates a new M instance from its JSON.parsed specification. This allows M to be defined within JSON schedules.

Construction

Model(definition; bindings = Dict{Symbol, Any}())

Construct a Model subtype from definition.

If definition is not a Specification, it will first be converted to one. Then, if definition is a Template, it will be expanded to (and asserted to actually be) a Model, and otherwise it will be put into a Schedule.

The bindings will be used for template expansion (if it contains such references) or bound to the Schedule on construction.

Invocation

(f!::Model)(x, Δt::Float64; arguments...) = error("unimplemented")

Advance the simulation state x by at most Δt time units according to the dynamics defined by f!.

Subtypes <: Model{State} must define specialized methods for this function, one for each type of x they support, but at least for State.

Methods must advance the simulation state and return it, and they may do so as an arbitrary type. They may arbitrarily modify x in the process, and they may return (i.e. alias) it. Methods may assume that Δt ≥ 0.0 and may accept additional keyword arguments to control their behavior.

GeneRegulatorySystems.Models.Instant — Type

Instant{State} <: Model{State}

Abstract supertype of instant models that ignore their time budget parameter (Δt) and never progress on the time axis.

GeneRegulatorySystems.Models.Wrapped — Type

Wrapped{State} <: Model{State}

Wraps a Model together with an arbitrary definition object.

Invoking a Wrapped model will just forward the call to the contained model.

This is used to annotate any Model with the definition that produced it, which can either be a specification or model-specific definition object (such as Models.V1.Definition) or a Scheduling.Locator indicating where the model was defined within a Schedule. Because model can be Wrapped itself, this mechanism can be used to annotate Models with a chain of provenance, and the intermediate definitions produced during model construction can be recreated using Scheduling.reify.